make config globally accessible via import; refactor htsohm algorithm code out out ./hts.py
